Tidy up Swift scenarios
[full ci]
