+++ features/fixtures/shared/scenarios/AppAndDeviceAttributesUnhandledExceptionAfterLaunchScenario.swift
@@ -1,0 +1,12 @@
+class AppAndDeviceAttributesUnhandledExceptionAfterLaunchScenario: Scenario {
+
+    override func startBugsnag() {
+        config.autoTrackSessions = false
+        super.startBugsnag()
+    }
+
+    override func run() {
+        Bugsnag.markLaunchCompleted()
+        NSException(name: .genericException, reason: "isLaunching should be false after `Bugsnag.markLaunchCompleted()`").raise()
+    }
+}
