+++ features/fixtures/shared/scenarios/AppAndDeviceAttributesUnhandledExceptionDuringLaunchScenario.swift
@@ -1,0 +1,11 @@
+class AppAndDeviceAttributesUnhandledExceptionDuringLaunchScenario: Scenario {
+
+    override func startBugsnag() {
+        config.autoTrackSessions = false
+        super.startBugsnag()
+    }
+
+    override func run() {
+        NSException(name: .genericException, reason: "isLaunching should be true").raise()
+    }
+}
