+++ features/fixtures/shared/scenarios/AppHangDefaultConfigScenario.swift
@@ -1,0 +1,9 @@
+class AppHangDefaultConfigScenario: Scenario {
+    
+    override func run() {
+        let timeInterval: TimeInterval = 5
+        NSLog("Simulating an app hang of \(timeInterval) seconds...")
+        Thread.sleep(forTimeInterval: timeInterval)
+        NSLog("Finished sleeping")
+    }
+}
