+++ features/fixtures/shared/scenarios/AppHangDidEnterBackgroundScenario.swift
@@ -1,0 +1,11 @@
+import UIKit
+
+class AppHangDidEnterBackgroundScenario: Scenario {
+    
+    override func run() {
+        NotificationCenter.default.addObserver(forName: UIApplication.didEnterBackgroundNotification, object: nil, queue: nil) {
+            NSLog("Recevied \($0.name), now hanging indefinitely...")
+            while true {}
+        }
+    }
+}
