+++ features/fixtures/shared/scenarios/AppHangDisabledScenario.swift
@@ -1,0 +1,14 @@
+class AppHangDisabledScenario: Scenario {
+
+    override func startBugsnag() {
+        config.enabledErrorTypes.appHangs = false
+        super.startBugsnag()
+    }
+    
+    override func run() {
+        let timeInterval: TimeInterval = 5
+        NSLog("Simulating an app hang of \(timeInterval) seconds...")
+        Thread.sleep(forTimeInterval: timeInterval)
+        NSLog("Finished sleeping")
+    }
+}
