+++ features/fixtures/shared/scenarios/EnabledReleaseStageManualSessionScenario.swift
@@ -1,0 +1,13 @@
+class EnabledReleaseStageManualSessionScenario : Scenario {
+
+    override func startBugsnag() {
+        self.config.autoTrackSessions = false
+        self.config.releaseStage = "prod"
+        self.config.enabledReleaseStages = ["dev", "prod"]
+        super.startBugsnag()
+    }
+
+    override func run() {
+        Bugsnag.startSession()
+    }
+}
