+++ features/fixtures/shared/scenarios/HandledErrorValidReleaseStageScenario.swift
@@ -1,0 +1,17 @@
+class MagicError : NSError {}
+
+class HandledErrorValidReleaseStageScenario : Scenario {
+
+    override func startBugsnag() {
+        self.config.autoTrackSessions = false;
+        self.config.releaseStage = "prod"
+        self.config.enabledReleaseStages = ["dev", "prod"]
+        super.startBugsnag()
+    }
+
+    override func run() {
+        Bugsnag.notifyError(MagicError(domain: "com.example",
+                                       code: 43,
+                                       userInfo: [NSLocalizedDescriptionKey: "incoming!"]))
+    }
+}
