+++ features/fixtures/shared/scenarios/SendLaunchCrashesSynchronouslyFalseScenario.swift
@@ -1,0 +1,7 @@
+class SendLaunchCrashesSynchronouslyFalseScenario: SendLaunchCrashesSynchronouslyScenario {
+    
+    override func startBugsnag() {
+        config.sendLaunchCrashesSynchronously = false
+        super.startBugsnag()
+    }
+}
