+++ features/fixtures/shared/scenarios/SendLaunchCrashesSynchronouslyLaunchCompletedScenario.swift
@@ -1,0 +1,10 @@
+class SendLaunchCrashesSynchronouslyLaunchCompletedScenario: SendLaunchCrashesSynchronouslyScenario {
+    
+    override func run() {
+        if eventMode != "report" {
+            NSLog(">>> Calling markLaunchCompleted()")
+            Bugsnag.markLaunchCompleted()
+        }
+        super.run()
+    }
+}
