+++ features/fixtures/shared/scenarios/UnhandledErrorInvalidReleaseStageScenario.swift
@@ -1,0 +1,13 @@
+class UnhandledErrorInvalidReleaseStageScenario : Scenario {
+
+    override func startBugsnag() {
+        self.config.autoTrackSessions = false;
+        self.config.releaseStage = "dev"
+        self.config.enabledReleaseStages = ["prod"]
+        super.startBugsnag()
+    }
+
+    override func run() {
+        abort();
+    }
+}
