+++ features/fixtures/shared/scenarios/UnhandledErrorValidReleaseStageScenario.swift
@@ -1,0 +1,13 @@
+class UnhandledErrorValidReleaseStageScenario : Scenario {
+
+    override func startBugsnag() {
+        self.config.autoTrackSessions = false;
+        self.config.releaseStage = "prod"
+        self.config.enabledReleaseStages = ["dev", "prod"]
+        super.startBugsnag()
+    }
+
+    override func run() {
+        abort();
+    }
+}
